recaptcha_FA/requirements.txt
azure-functions
google-cloud-recaptcha-enterprise